Progressing on the label bug.
Now the label is correctly parsed from the ASM column.
The rest of the code still needs to be able to know it has been. There exists a table of label (labelTable), look into how it's built and how loading a ROM works. Does it fill the table before assembling?