Clear and create labelTable on assembling
We need to clear the labelTable on assembling
otherwise the table is corrupted by previous labels.
IMO, using a "label1", assembling, removing it,
and then "beq 0,0,label1" won't return an error without this change,
since "label1" will still be in the labelTable.

To create the labelTable, it fetches labels from:
 - the command (label: ...)
 - the address column
Labels containing only numbers are ignored to avoid address collisions.