Use instruction code to detect current instruction
We need to know the current instruction to enable/disable components.
Instead of using a heuristic string search on the string command,
we can use the instruction code generated when assembling.
After assembling, we should not rely on any string anymore,
but the instruction code.
Parsing user inputs is hard. Do it 1!.

Related to #14, but the other simulators might still be affected.